pyproject.toml
[tool.poetry]
name = "drf-anifi"
version = "0.1.0"
description = ""
authors = ["Vichaya Raksakunpanich <toonbcc@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
psycopg2-binary = "^2.9.3"
whitenoise = {extras = ["brotli"], version = "^6.2.0"}
gunicorn = "^20.1.0"
python-dotenv = "^0.20.0"
djangorestframework = "^3.13.1"
Markdown = "^3.3.7"
django-filter = "^21.1"
django-cors-headers = "^3.13.0"
drf-spectacular = "^0.22.1"
djangorestframework-simplejwt = "^5.2.0"
web3 = "^5.29.2"
sentry-sdk = "^1.6.0"
hiredis = "^2.0.0"
redis = "^4.3.4"
factory-boy = "^3.2.1"
django-debug-toolbar = "^3.5.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
